Worker earning reputation through task completion

+++ contracts/ERC20Extended.sol
@@ -6,5 +6,5 @@ import "../lib/dappsys/erc20.sol";
 
 
 contract ERC20Extended is ERC20 {
-  function mint(uint128 wad) public;
+  function mint(uint wad) public;
 }
